Update video FPS attribute to float32
Turns out Google Photos will not always return an integer for this.

+++ internal/photoslibrary/photoslibrary.go
@@ -56,7 +56,7 @@ type Photo struct {
 type Video struct {
 	CameraMake  string
 	CameraModel string
-	Fps         int
+	Fps         float32
 	Status      string
 }
